+++ php/classes/circle.inc.php
@@ -251,14 +251,14 @@ class circle extends zophTable {
         $qry->addParam(new param(":circleid", (int) $this->getId(), PDO::PARAM_INT));
 
         if (!user::getCurrent()->canSeeAllPhotos()) {
-            $allowed=person::getAllPeopleAndPhotoGraphers();
+            $allowed=person::getAllPeopleAndPhotographers();
             $ids=array();
             foreach ($allowed as $person) {
                 $ids[]=$person->getId();
             }
             $param=new param(":peopledIds", $ids, PDO::PARAM_INT);
             $qry->addParam($param);
-            $qry->addClause(clause::InClause("person_id", $param, "AND"));
+            $qry->addClause(clause::InClause("person_id", $param), "AND");
         }
 
         return person::getRecordsFromQuery($qry);