+++ docs/CHANGELOG.md
@@ -6,7 +6,8 @@
 #### Refactor ####
 
 * Changing `and` and `or` into `&&` and `||` for readability and consistency
-
+* Changed capitalization of exception classes for consistency
+* Several fixes in namespace, capitalization and documentation
 
 ## Zoph 0.9.8 ##
 ### 02 march 2018 ###